Develop to Release

+++ app/Http/Controllers/UserController.php
@@ -3,8 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Enums\UserRole;
+use App\Models\Customer;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 
 /**
@@ -41,8 +43,9 @@ class UserController extends Controller
         $user->is_can_access_customer = 0;
 
         $userRoles = UserRole::asSelectArray();
+        $customerAccounts = Customer::all();
 
-        return view('user.create', compact('user', 'userRoles'));
+        return view('user.create', compact('user', 'userRoles', 'customerAccounts'));
     }
 
     /**
@@ -71,6 +74,11 @@ class UserController extends Controller
      */
     public function show($id)
     {
+        if (config('auth.super_admin_user_id') == $id && Auth::user()->id != $id) {
+            return redirect()->route('users.index')
+                ->with('success', 'Can not show super admin');
+        }
+
         $user = User::find($id);
 
         return view('user.show', compact('user'));
@@ -84,11 +92,17 @@ class UserController extends Controller
      */
     public function edit($id)
     {
+        if (config('auth.super_admin_user_id') == $id && Auth::user()->id != $id) {
+            return redirect()->route('users.index')
+                ->with('success', 'Can not edit super admin');
+        }
+
         $user = User::find($id);
 
         $userRoles = UserRole::asSelectArray();
+        $customerAccounts = Customer::all();
 
-        return view('user.edit', compact('user', 'userRoles'));
+        return view('user.edit', compact('user', 'userRoles', 'customerAccounts'));
     }
 
     /**
@@ -100,6 +114,11 @@ class UserController extends Controller
      */
     public function update(Request $request, User $user)
     {
+        if (config('auth.super_admin_user_id') == $user->id && Auth::user()->id != $user->id) {
+            return redirect()->route('users.index')
+                ->with('success', 'Can not edit super admin');
+        }
+
         request()->validate(User::$onUpdateRules);
 
         $data = $request->all();
@@ -122,6 +141,11 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        if (config('auth.super_admin_user_id') == $id && Auth::user()->id != $id) {
+            return redirect()->route('users.index')
+                ->with('success', 'Can not delete super admin');
+        }
+
         $user = User::find($id)->delete();
 
         return redirect()->route('users.index')
